Reduced duplication of html transform logic
refs https://github.com/TryGhost/Team/issues/467

- absolute and relative html transformations had a large amount of duplicated code
- by pulling the duplicated code out into a new generic function that can accept specific transformation functions it can be re-used for the new transform-ready output too
